Change u: with v in the cedict file
This was made as converting 'plain' pinyin (ni3 hao3) to 'accented'
pinyin (nǐ hǎo) was causing problems with u:. We could change the
way the accented pinyin is calculated, but this was a quicker fix.

In the future I expect to fix it so it is easy to update the cedict
file directly (without having to replace all 'u:'s for 'v's).